Crate prelude

Re-exports

pub use ahash::HashMapExt as _;

pub use ahash::HashSetExt as _;

pub use itermore::IterArrayChunks as _;

pub use itermore::IterArrayCombinations as _;

pub use itermore::IterArrayWindows as _;

pub use itermore::IterCartesianProduct as _;

pub use itermore::IterCircularArrayWindows as _;

pub use itermore::IterCollectArray as _;

pub use itermore::IterCombinations as _;

pub use itermore::IterCombinationsWithReps as _;

pub use itermore::IterMinMax as _;

pub use itermore::IterNextChunk as _;

pub use itermore::IterSorted as _;

pub use then::Some as _;

Modules

cmp

Utilities for comparing and ordering values.

iter

Composable external iteration.

mem

Basic functions for dealing with memory.

Macros

cartesian_product

Returns an iterator over the cartesian product of the element sets of multiple iterators (up to 12).

include_input

regex

Creates a static regex and returns a reference to it.

vector

A macro for composing vectors.

vectors

Structs

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

BinaryHeap

A priority queue implemented with a binary heap.

Matrix

Represents a matrix with constant M rows and constant N columns.

Reverse

A helper struct for reverse ordering.

VecDeque

A double-ended queue implemented with a growable ring buffer.

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Ordering

An Ordering is the result of a comparison between two values.

Traits

ItertoolsShim

Functions

gcd

Returns the greatest common divisor of two numbers.

lcm

Returns the least common multiple of two numbers.

max

Compares and returns the maximum of two values.

min

Compares and returns the minimum of two values.

parse_map

Parses a 2D map into a data structure of type M.

parse_map_set

Parses a 2D map that uses ‘.’ for spaces and ‘#’ for walls as a set of the the ‘#’ points.

Type Aliases

HashMap

Type alias for HashMap<K, V, ahash::RandomState>

HashSet

Type alias for HashSet<K, ahash::RandomState>

Vector

A matrix with one column and M rows.

Vector2

Vector3

Vector4