Crate prelude

Prelude is designed to be imported with use prelude::*.

Contains reexports from std & basic often needed stuff.

Re-exports

pub use num_integer::Integer;

pub use num_traits::clamp;

pub use num_traits::Float;

pub use num_traits::Num;

pub use std::borrow::Cow;

pub use std::cell::Cell;

pub use std::cell::Ref;

pub use std::cell::RefCell;

pub use std::cell::RefMut;

pub use std::cmp::Eq;

pub use std::cmp::Ord;

pub use std::cmp::PartialEq;

pub use std::cmp::PartialOrd;

pub use std::collections::BTreeMap;

pub use std::collections::BTreeSet;

pub use std::collections::HashMap;

pub use std::collections::HashSet;

pub use std::ffi::CStr;

pub use std::ffi::CString;

pub use std::fmt::Debug;

pub use std::fmt::Display;

pub use std::fmt::Formatter;

pub use std::marker::PhantomData;

pub use std::mem;

pub use std::ops::Add;

pub use std::ops::AddAssign;

pub use std::ops::Div;

pub use std::ops::DivAssign;

pub use std::ops::Mul;

pub use std::ops::MulAssign;

pub use std::ops::Neg;

pub use std::ops::Sub;

pub use std::ops::SubAssign;

pub use std::ops::Deref;

pub use std::ops::DerefMut;

pub use std::ops::Range;

pub use std::ops::RangeFrom;

pub use std::ops::RangeFull;

pub use std::ops::RangeTo;

pub use std::ops::Index;

pub use std::ops::IndexMut;

pub use std::os::raw::c_char;

pub use std::os::raw::c_double;

pub use std::os::raw::c_float;

pub use std::os::raw::c_int;

pub use std::os::raw::c_long;

pub use std::os::raw::c_short;

pub use std::os::raw::c_ulong;

pub use std::os::raw::c_ushort;

pub use std::os::raw::c_void;

pub use std::rc::Rc;

pub use std::sync::Arc;

pub use std::sync::Mutex;

pub use std::thread;

Structs

ACell
Atomic cell. Just like Cell, but thread-safe.

CustomRange

Traits

RangeArgument
Temporary re-implemenation of RangeArgument from stdlib while waiting for it to become stable

StableFn

StableFnMut

StableFnOnce

Functions

default
Shortcut for Default::default.

get_slice

max
Get maximum of two values.

min
Get minimum of two values.

min_max
Sort two values.